requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.2=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
filelock=3.9.0=pypi_0
fonttools=4.44.0=pypi_0
fsspec=2023.4.0=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
grpcio=1.59.2=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_2
libgomp=13.2.0=h807b86a_2
libnsl=2.0.1=hd590300_0
libsqlite=3.44.0=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=hd590300_5
lightning-utilities=0.9.0=pypi_0
markdown=3.5.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.8.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h59595ed_2
networkx=3.0=pypi_0
numpy=1.24.1=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.1.4=hd590300_0
packaging=23.2=pypi_0
pandas=2.1.2=pypi_0
pillow=9.3.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
prettytable=3.9.0=pypi_0
protobuf=4.23.4=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.9.13=h2660328_0_cpython
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h8228510_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
seaborn=0.13.0=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
sqlite=3.44.0=h2c6b66d_0
sympy=1.12=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch=2.1.0+cu118=pypi_0
torchaudio=2.1.0+cu118=pypi_0
torcheval=0.0.7=pypi_0
torchmetrics=1.2.0=pypi_0
torchvision=0.16.0+cu118=pypi_0
tqdm=4.66.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.13=pypi_0
wcwidth=0.2.9=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.3=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zipp=3.17.0=pypi_0
